requirements.txt
# Core Python libraries
pandas>=1.3.0          # Data manipulation
torch>=2.0.0           # PyTorch for deep learning
transformers>=4.29.0   # Hugging Face Transformers for zero-shot classification
scikit-learn>=1.0      # K-Means clustering and other utilities
requests>=2.28.0       # Scraping and HTTP requests
openai>=1.58.0         # GPT-4 API integration
python-dotenv>=0.21.0  # Dotenv for config files
tqdm>=4.62.0           # Progress bars for loops (optional but helpful)
requests>=2.26.0       # Utils for scraping and retry
urllib3>=1.26.0        # Utils for scraping and retry